Run InputDate E2E tests on WebAssembly only. Fixes #35018, #34884
